Fix the delete modal

+++ hapus.php
@@ -1,10 +1,16 @@
 <?php
 require 'koneksi.php';
 
-$id =  $_GET['id'];
+
+
+if(isset($_GET['yakin'])){
+$id =  $_GET['id_to_delete'];
 mysqli_query($koneksi, "DELETE FROM plus_poin WHERE id =  '$id'");
 echo "<script>window.location.href = 'pluspoint.php'</script>";
 exit;
+}else{
+    return false;
+}
 
 
 ?>
